unsigned int for age

+++ internal/datastructures/sql/sql.go
@@ -6,7 +6,7 @@ type PatientInfo struct {
 	UserID      *string    `json:"userID,omitempty" required:"false"`
 	FirstName   *string    `json:"firstName" doc:"First name of the user"`
 	LastName    *string    `json:"lastName" doc:"Last name of the user" required:"false"`
-	Age         *int       `json:"age" doc:"Age of the user"`
+	Age         *uint16    `json:"age" doc:"Age of the user"`
 	BloodGroup  *string    `json:"bloodGroup" doc:"Blood Group of the user" required:"false"`
 	PhoneNumber *string    `json:"phoneNumber" doc:"Phone number of the user" minLength:"10" maxLength:"12"`
 	Address     *string    `json:"address" doc:"Address of the user" required:"false"`
